data-glossar: fallback to current web
Because we now allow multiple glossaries it is no longer clear in which
web to look for an entry. Some older view-templates do not include the
web name in their data-glossar attribute. I will now fallback to the
current web in that case. If the entry is located in another web, the
view-template needs to be updated.

Release-Note(fix, admin) {
  Fix glossary portals not showing the correct popups when multiple
  glossaries are defined.
  [...]
  In some cases the plugin would attempt to fetch the popup from the
  wrong web when multiple glossaries are defined. This happened only on
  the portal pages (WebHome).
}

Reviewed on merge-day.

Ticket: #6249